bit_lib moved from lfrfid to standalone lib
