+++ applications/main/nfc/plugins/supported_cards/umarsh.c
@@ -30,7 +30,7 @@
 #include <flipper_application/flipper_application.h>
 
 #include <nfc/nfc_device.h>
-#include <lfrfid/tools/bit_lib.h>
+#include <bit_lib/bit_lib.h>
 #include <nfc/protocols/mf_classic/mf_classic_poller_sync.h>
 
 #include <furi_hal_rtc.h>
