+++ applications/main/nfc/plugins/supported_cards/washcity.c
@@ -24,7 +24,7 @@
 #include <flipper_application/flipper_application.h>
 
 #include <nfc/nfc_device.h>
-#include <lfrfid/tools/bit_lib.h>
+#include <bit_lib/bit_lib.h>
 #include <nfc/protocols/mf_classic/mf_classic_poller_sync.h>
 
 #define TAG "WashCity"
@@ -141,7 +141,8 @@ static bool washcity_parse(const NfcDevice* device, FuriString* parsed_data) {
         const MfClassicSectorTrailer* sec_tr =
             mf_classic_get_sector_trailer_by_sector(data, ticket_sector_number);
 
-        const uint64_t key = nfc_util_bytes2num(sec_tr->key_a.data, COUNT_OF(sec_tr->key_a.data));
+        const uint64_t key =
+            bit_lib_bytes_to_num_be(sec_tr->key_a.data, COUNT_OF(sec_tr->key_a.data));
         if(key != washcity_1k_keys[ticket_sector_number].a) break;
 
         // Parse data
@@ -151,7 +152,7 @@ static bool washcity_parse(const NfcDevice* device, FuriString* parsed_data) {
         const uint8_t* block_start_ptr =
             &data->block[start_block_num + ticket_block_number].data[0];
 
-        uint32_t balance = nfc_util_bytes2num(block_start_ptr + 2, 2);
+        uint32_t balance = bit_lib_bytes_to_num_be(block_start_ptr + 2, 2);
 
         uint32_t balance_eur = balance / 100;
         uint8_t balance_cents = balance % 100;
@@ -160,7 +161,7 @@ static bool washcity_parse(const NfcDevice* device, FuriString* parsed_data) {
         const uint8_t* uid = mf_classic_get_uid(data, &uid_len);
 
         // Card Number is printed in HEX (equal to UID)
-        uint64_t card_number = nfc_util_bytes2num(uid, uid_len);
+        uint64_t card_number = bit_lib_bytes_to_num_be(uid, uid_len);
 
         furi_string_printf(
             parsed_data,
