+++ lib/SConscript
@@ -41,6 +41,7 @@ libs = env.BuildModules(
         "nanopb",
         "update_util",
         "heatshrink",
+        "bit_lib",
     ],
 )
