+++ lib/bit_lib/bit_lib.h
