+++ lib/lfrfid/lfrfid_worker_modes.c
@@ -5,7 +5,7 @@
 #include <toolbox/pulse_protocols/pulse_glue.h>
 #include <toolbox/buffer_stream.h>
 #include "tools/varint_pair.h"
-#include "tools/bit_lib.h"
+#include <lib/bit_lib/bit_lib.h>
 
 #define TAG "LfRfidWorker"
