+++ lib/lfrfid/protocols/protocol_fdx_b.c
@@ -2,7 +2,7 @@
 #include "toolbox/level_duration.h"
 #include "protocol_fdx_b.h"
 #include <toolbox/manchester_decoder.h>
-#include <lfrfid/tools/bit_lib.h>
+#include <bit_lib/bit_lib.h>
 #include "lfrfid_protocols.h"
 #include <furi_hal_rtc.h>
