+++ lib/lfrfid/protocols/protocol_jablotron.c
@@ -2,7 +2,7 @@
 #include "toolbox/level_duration.h"
 #include "protocol_jablotron.h"
 #include <toolbox/manchester_decoder.h>
-#include <lfrfid/tools/bit_lib.h>
+#include <bit_lib/bit_lib.h>
 #include "lfrfid_protocols.h"
 
 #define JABLOTRON_ENCODED_BIT_SIZE (64)
