+++ lib/lfrfid/protocols/protocol_viking.c
@@ -1,7 +1,7 @@
 #include <furi.h>
 #include <toolbox/protocols/protocol.h>
 #include <toolbox/manchester_decoder.h>
-#include <lfrfid/tools/bit_lib.h>
+#include <bit_lib/bit_lib.h>
 #include "lfrfid_protocols.h"
 
 #define VIKING_CLOCK_PER_BIT (32)
