+++ lib/nfc/protocols/mf_classic/crypto1.c
@@ -1,7 +1,7 @@
 #include "crypto1.h"
 
 #include <lib/nfc/helpers/nfc_util.h>
-#include <lib/lfrfid/tools/bit_lib.h>
+#include <lib/bit_lib/bit_lib.h>
 #include <furi.h>
 
 // Algorithm from https://github.com/RfidResearchGroup/proxmark3.git
