+++ lib/nfc/protocols/mf_ultralight/mf_ultralight_poller_i.h
@@ -2,7 +2,7 @@
 
 #include "mf_ultralight_poller.h"
 #include <lib/nfc/protocols/iso14443_3a/iso14443_3a_poller_i.h>
-#include <lib/lfrfid/tools/bit_lib.h>
+#include <lib/bit_lib/bit_lib.h>
 
 #ifdef __cplusplus
 extern "C" {
