+++ targets/f7/target.json
@@ -47,6 +47,7 @@
         "update_util",
         "heatshrink",
         "flipperformat",
-        "flipper7"
+        "flipper7",
+        "bit_lib"
     ]
 }
